change service by router
