+++ src/main/resources/spring/camel-context.xml
@@ -237,7 +237,7 @@
                     <to uri="direct:to-sortBean2"/>
                     <to uri="direct:to-workloadInventoryFilterBean"/>
                     <bean ref="workloadInventoryReportService" method="findByAnalysisOwnerAndAnalysisId(${header.${type:org.jboss.xavier.integrations.route.RouteBuilderExceptionHandler.USERNAME}}, ${header.id}, ${header.pageBean}, ${header.sortBean}, ${header.workloadInventoryFilterBean})" />
-                    <bean ref="openApiService" method="buildResponse(${headers}, ${body}, null)" />
+                    <to uri="direct:to-openapi-paginationResponse" />
                 </route>
             </get>
             <get uri="/{id}/workload-inventory/filtered-csv" produces="text/csv">
